Samler hjelpefunksjoner rundt tidsrom, dvs fra-og-med og til-og-med ett sted
