+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/tid/Tidspunkt.kt
@@ -44,31 +44,3 @@ abstract class Tidspunkt<T : Tidsenhet> internal constructor(
         }
     }
 }
-
-fun <T : Tidsenhet> størsteAv(t1: Tidspunkt<T>, t2: Tidspunkt<T>): Tidspunkt<T> =
-    if (t1.erUendeligLengeTil() && t2.erEndelig() && t1.somEndelig() <= t2) {
-        t2.neste().somUendeligLengeTil()
-    } else if (t2.erUendeligLengeTil() && t1.erEndelig() && t2.somEndelig() <= t1) {
-        t1.neste().somUendeligLengeTil()
-    } else if (t1.erUendeligLengeTil() || t2.erUendeligLengeTil()) {
-        maxOf(t1.somEndelig(), t2.somEndelig()).somUendeligLengeTil()
-    } else {
-        maxOf(t1, t2)
-    }
-
-fun <T : Tidsenhet> minsteAv(t1: Tidspunkt<T>, t2: Tidspunkt<T>): Tidspunkt<T> =
-    if (t1.erUendeligLengeSiden() && t2.erEndelig() && t1.somEndelig() >= t2) {
-        t2.forrige().somUendeligLengeSiden()
-    } else if (t2.erUendeligLengeSiden() && t1.erEndelig() && t2.somEndelig() >= t1) {
-        t1.forrige().somUendeligLengeSiden()
-    } else if (t1.erUendeligLengeSiden() || t2.erUendeligLengeSiden()) {
-        minOf(t1.somEndelig(), t2.somEndelig()).somUendeligLengeSiden()
-    } else {
-        minOf(t1, t2)
-    }
-
-fun <T : Tidsenhet> Iterable<Tidspunkt<T>>.størsteEllerNull() =
-    this.reduceOrNull { acc, neste -> størsteAv(acc, neste) }
-
-fun <T : Tidsenhet> Iterable<Tidspunkt<T>>.minsteEllerNull() =
-    this.reduceOrNull { acc, neste -> minsteAv(acc, neste) }
