+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/tidslinje/TidsromTest.kt
@@ -1,9 +1,7 @@
 package no.nav.familie.ba.sak.kjerne.tidslinje
 
-import no.nav.familie.ba.sak.kjerne.tidslinje.tid.minsteAv
 import no.nav.familie.ba.sak.kjerne.tidslinje.tid.somUendeligLengeSiden
 import no.nav.familie.ba.sak.kjerne.tidslinje.tid.somUendeligLengeTil
-import no.nav.familie.ba.sak.kjerne.tidslinje.tid.størsteAv
 import no.nav.familie.ba.sak.kjerne.tidslinje.util.apr
 import no.nav.familie.ba.sak.kjerne.tidslinje.util.des
 import no.nav.familie.ba.sak.kjerne.tidslinje.util.feb
@@ -15,7 +13,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotEquals
 import org.junit.jupiter.api.Test
 
-class TidspunktTest {
+class TidsromTest {
     @Test
     fun testStørsteAv() {
         assertEquals(
